Allow simple comparators for filtering
Currently, it is possible to filter for values greater or smaller than
a threshold with 'X...' or '...X'. This is not the most intuitive
notation, furthermore selecting only one value requires a lot of typing
('X...X'). This commit implements an alternative way of filtering
with '>=X', '<=X' and '=X'.