Fix Crosstalk bug due to bad variable rename
I had changed instance.ctselect to instance.ctselectHandle, but due to bad Find criteria I
missed some of the references. The effect of this bug was that selecting rows in the DT
would cause the same table to respond to the selection as if it was originating elsewhere,
so all but the selected rows would be hidden.