Always include Crosstalk dependencies
The JavaScript bindings are structured in such a way that Crosstalk
dependencies are always needed. This results in an extra ~19KB
(minified) JS dependency. We could definitely remove this dependency by
being more careful in the JS bindings not to touch crosstalk.* until a
group is passed in, but in the meantime this fixes fatal errors.